ProtectedAccess.php simplify
